temp genstmt
